data/requirements.txt
torch>=2.0.0
numpy
pandas
pyyaml
scikit-learn
matplotlib
gym
